requirements.txt
# Core orchestration
langgraph==0.2.28
langchain-core==0.3.15

# LLM
google-generativeai==0.8.3
ollama==0.3.3

# Browser automation
playwright==1.47.0

# HTML parsing
beautifulsoup4==4.12.3
lxml==5.3.0

# Resume editing
python-docx==1.1.2

# Config
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Scheduler
APScheduler==3.10.4

# Dashboard
streamlit==1.39.0
pandas==2.2.3
plotly==5.24.1
altair==5.4.1

# Utilities
# difflib2==0.2.0
requests==2.32.3
